requirements.txt
selenium
webdriver-manager
schedule
python-dotenv
